Chunked transfer encoding added.
Chunked transfer encoding allows to send the body in multiple packages.
Said packages are marked by their size in hex, a CRLF and the content.
The last package must be followed by an empty package marked as such :
0CRLFCRLF
I think we can do better and store the out file of get method in RC, and read it
package_size by package_size