Removed upload path and transformed allowed methods into a set rather than a map
Upload_path can be handled using the following in the config file : methods /post; OR not_methods post;
Methods are no longer stored in a map containing whether or not they are allowed.
If a method is in the set, it is allowed. An unauthorized method shall be removed.
Added the header 'Accept' containing allowed methods in case of a 405 error