Full account management :)
Cookies are no longer created by JS and are fully managed by the server with Set-Cookie header.
Disconnecting is made through a DELETE method, and not a POST.
Session management is done through a .session file at the root of the repo.
This file associates an auth key to an account name (stored in .usr)
Sessions are added every time, but are only removed at the end due to file management crap (might need to be improved).
Anyway, that's done